Refactor icon file structure
