--- frontend/libs/studio-icons/src/components/Accordion/index.ts
@@ -1,1 +1,0 @@
-export { Accordion } from './Accordion';
