--- frontend/libs/studio-icons/src/components/Group/index.ts
@@ -1,1 +1,0 @@
-export { Group } from './Group';
