--- frontend/libs/studio-icons/src/components/Likert/index.ts
@@ -1,1 +1,0 @@
-export { Likert } from './Likert';
