--- frontend/libs/studio-icons/src/components/LongText/index.ts
@@ -1,1 +1,0 @@
-export { LongText } from './LongText';
