--- frontend/libs/studio-icons/src/components/NavBar/index.ts
@@ -1,1 +1,0 @@
-export { NavBar } from './NavBar';
