--- frontend/libs/studio-icons/src/components/Paragraph/index.ts
@@ -1,1 +1,0 @@
-export { Paragraph } from './Paragraph';
