--- frontend/libs/studio-icons/src/components/RadioButton/index.ts
@@ -1,1 +1,0 @@
-export { RadioButton } from './RadioButton';
