--- frontend/libs/studio-icons/src/components/ShortText/index.ts
@@ -1,1 +1,0 @@
-export { ShortText } from './ShortText';
