--- frontend/libs/studio-icons/src/components/SvgTemplate/index.ts
@@ -1,1 +1,0 @@
-export { SvgTemplate } from './SvgTemplate';
