--- frontend/libs/studio-icons/src/components/Title/index.ts
@@ -1,1 +1,0 @@
-export { Title } from './Title';
