+++ frontend/libs/studio-icons/src/index.ts
@@ -1,5 +1,5 @@
 // Re-export all icons from @navikt/aksel-icons to make them available for consumers of studio/icons
 export * from '@navikt/aksel-icons';
 
-export * from './components';
-export * from './types/IconProps';
+export * from './react/icons';
+export * from './react/types';
