+++ frontend/libs/studio-icons/src/react/icons/Icons.test.tsx
