+++ frontend/libs/studio-icons/src/react/icons/index.ts
