+++ frontend/libs/studio-icons/src/react/types/IconProps.ts
