+++ frontend/libs/studio-icons/src/react/types/index.ts
@@ -1,0 +1,1 @@
+export type { IconProps } from './IconProps';
